pyproject.toml
[project]
name = "py-gitlab-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nicegui>=2.10.1",
    "orjson>=3.10.15",
    "python-gitlab>=5.3.1",
    "ruff>=0.9.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]

[tool.ruff]
line-length = 120
